chore(fields): Index.js has the new Field component
